refactor: satisfy release architecture budgets
