+++ style.css
@@ -12,6 +12,7 @@
 @import url("./styles/05-instantiation-dialog.css");
 @import url("./styles/06-layout-controls.css");
 @import url("./styles/07-properties.css");
+@import url("./styles/07-properties-pills.css");
 @import url("./styles/07-properties-images.css");
 @import url("./styles/08-event-log.css");
 @import url("./styles/09-script-console.css");
