+++ styles/07-properties-pills.css
@@ -1,0 +1,23 @@
+/* ── Properties status pills ──────────────────────────────── */
+
+.pill {
+    min-width: 18px;
+    height: 18px;
+    padding: 0 6px;
+    border-radius: 4px;
+    border: 0;
+    background: var(--neon);
+    color: var(--text-dark);
+    font-family: "Space Mono", "SFMono-Regular", Consolas, monospace;
+    font-size: 10px;
+    font-weight: 600;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.empty-state {
+    color: var(--text-muted);
+    font-size: 11px;
+    font-family: "Space Mono", "SFMono-Regular", Consolas, monospace;
+}
